+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,12 @@
+By contributing to Flying-Router, you agree to follow this code of conduct.
+
+To keep things short: Be nice and **DO NOT** spam, troll or harass someone.
+
+- **Trolling** includes to post comments to provoke someone else, or disrupting a discussion.
+- **Spam** includes promoting a product, advertise a job, flooding discussions with non related comments or unnecessary amounts of comments and do not send off-topic messages in a discussion, all this disrupts discussions and are not allowed
+- **Harassment** includes but is not limited to sexual language and imagery, deliberate intimidation, stalking, name-calling, unwelcome attention, libel, and any malicious hacking or social engineering. Flying-Router should be harassment-free for everyone regardless of the person behind the computer
+
+In the repository are no bots allowed without prior written permission from [HrPlanlos](https://github.com/HrPlanlos).
+
+After finding a violation against the code of conduct in the Flying-Router repository please reach out to [HrPlanlos](https://github.com/HrPlanlos).  
+If you can please make a screenshot with the offense and the url to help us find the offender faster, we will do everything we think is appropriate, up to banning the offender if the number of violations is too large or the or the act is that serious.
