+++ README.md
@@ -1,1 +1,42 @@
-# flying-router
+# [Flying Router](https://github.com/HrPlanlos/flying-router) &middot; [![License](https://img.shields.io/github/license/HrPlanlos/flying-router)]()
+
+Flying Router is a Router implementation.
+
+- **Dynamic**: Flying Router lets you decide what the type of your request and response is.
+- **Resource Efficient**: With the small size amd the focus on small environments, Flying Router doesn't waste resources on features needed only in enterprise environments.
+- **Simple**: The router has a very simple structure where you can define everything with only a few lines of code
+
+## Installation
+
+To install Flying Router you have to execute the following command:
+
+```sh
+$ npm i flying-router
+```
+
+See [Getting Started](https://github.com/HrPlanlos/flying-router/wiki/getting_started) to see an installation guide in more detail.
+
+(currently we only support npm, in the future we may add support for yarn)
+
+## Documentation
+
+You can find the Flying Router documentation in the [WIKI tab](https://github.com/HrPlanlos/flying-router/wiki) of the repo.
+
+If you find something missing in the wiki please feel free to add it, or open an issue and it will get added in near future.
+
+## Contributing
+
+The purpose of this repository is to develop a simple and lightweight router, we'd love to have **you** contribute to this project.  
+As for this reason we came up with some simple rules you need to follow if you want to contribute:
+
+### [Code of Conduct](https://github.com/HrPlanlos/flying-router/blob/main/CODE_OF_CONDUCT.md)
+
+We have come up with a Code of Conduct containing everything we expect from a project participant. Please read the full file before you start contributing, otherwise it could happen that we can't work with your idea.
+
+### How to contribute
+
+Everyone can contribute, first start witting an issue with your idea, after we read and thought about your idea we can tell you if it fits in this project. After you know it fits into the project you can start creating a fork and implementing your idea. After you are done adding your feature, create a pull request for your fork. We will review your code and if everything is fine its gonna get merged into the repo, otherwise if we find something that isn't right with your code, we will write a comment stating what you should change before we can merge your pull request.
+
+## License
+
+Flying Router uses [BSL 1.0](https://github.com/HrPlanlos/flying-router/blob/main/LICENSE)
